backend/requirements.txt
black==26.1.0
blinker==1.9.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==46.0.3
dataclasses==0.6
exceptiongroup==1.3.0
flake8==7.3.0
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.0
pytokens==0.4.0
requests==2.32.3
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.22.0
